requirements.txt
appdirs==1.4.3
appnope==0.1.0
arrow==0.6.0
decorator==4.0.11
Django==1.11.1
djangorestframework==3.6.3
facebook-sdk==2.0.0
Faker==0.7.12
ipdb==0.10.3
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
oauthlib==2.0.2
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.14.2
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.10.0
tapioca-facebook==0.6.1
tapioca-wrapper==1.4
traitlets==4.3.2
wcwidth==0.1.7
